+++ src/main/java/com/verdantartifice/primalmagick/common/research/ResearchEntries.java
@@ -13,6 +13,7 @@ import com.verdantartifice.primalmagick.common.research.keys.ResearchEntryKey;
 import com.verdantartifice.primalmagick.common.sources.SourceList;
 import com.verdantartifice.primalmagick.common.sources.Sources;
 import com.verdantartifice.primalmagick.common.stats.StatsPM;
+import com.verdantartifice.primalmagick.common.tags.ItemTagsForgeExt;
 import com.verdantartifice.primalmagick.common.tags.ItemTagsPM;
 
 import net.minecraft.core.Holder;
@@ -1678,7 +1679,7 @@ public class ResearchEntries {
                 .build());
         register(context, DISSOLUTION_CHAMBER, key -> ResearchEntry.builder(key).discipline(discipline).tier(ResearchTier.MASTER).icon(ItemsPM.DISSOLUTION_CHAMBER.get())
                 .parent(MASTER_MAGITECH).parent(MASTER_MANAWEAVING).parent(EARTHSHATTER_HAMMER)
-                .stage().requiredItem(ItemsPM.IRON_GRIT.get(), 20).requiredItem(ItemsPM.COPPER_GRIT.get(), 20).requiredItem(ItemsPM.GOLD_GRIT.get(), 10)
+                .stage().requiredItem(ItemTagsForgeExt.DUSTS_IRON, 20).requiredItem(ItemTagsForgeExt.DUSTS_COPPER, 20).requiredItem(ItemTagsForgeExt.DUSTS_GOLD, 10)
                         .requiredTheories(2).end()
                 .stage().attunement(Sources.EARTH, 3).recipe(ItemsPM.DISSOLUTION_CHAMBER.get()).recipe("iron_grit_from_dissolving_ore")
                         .recipe("iron_grit_from_dissolving_raw_metal").recipe("gold_grit_from_dissolving_ore")