+++ changelog.md
@@ -11,6 +11,7 @@ v5.0.1-beta
 * Made the Shear spell payload's research entry no longer require shearing things as a prerequisite
 * Removed the shear use statistic from mod tracking
 * Fixed Dissolution Chamber research requirements to use metal dust tags instead of specific items
+* Updated the Conjure Animal spell to allow summoning armadillos
 
 v5.0.0-beta
 -----------